requirements.txt
django==3.0
djangorestframework==3.11.0
googlemaps==3.1.4
requests==2.22.0
gunicorn==19.9.0
